+++ src/core/profile-provider/profile-provider.test.ts
@@ -4,7 +4,6 @@ import type {
   SecurityValues,
 } from '@superfaceai/ast';
 import { HttpScheme, SecurityType } from '@superfaceai/ast';
-import { mocked } from 'ts-jest/utils';
 
 import type { IFileSystem } from '../../interfaces';
 import { err, ok } from '../../lib';
@@ -151,7 +150,52 @@ describe('profile provider', () => {
         const mockProviderJsonWithParameters = mockProviderJson({
           name: 'test',
         });
-        mocked(fetchBind).mockResolvedValue({
+        // jest.mocked(fetchBind).mockResolvedValue({
+        //   services: [{ id: 'test-service', baseUrl: 'service/base/url' }],
+        //   securitySchemes: [
+        //     {
+        //       type: SecurityType.HTTP,
+        //       id: 'basic',
+        //       scheme: HttpScheme.BASIC,
+        //     },
+        //     {
+        //       id: 'api',
+        //       type: SecurityType.APIKEY,
+        //       in: ApiKeyPlacement.HEADER,
+        //       name: 'Authorization',
+        //     },
+        //     {
+        //       id: 'bearer',
+        //       type: SecurityType.HTTP,
+        //       scheme: HttpScheme.BEARER,
+        //       bearerFormat: 'some',
+        //     },
+        //     {
+        //       id: 'digest',
+        //       type: SecurityType.HTTP,
+        //       scheme: HttpScheme.DIGEST,
+        //     },
+        //   ],
+        //   defaultService: 'test-service',
+        //   parameters: [
+        //     {
+        //       name: 'first',
+        //       description: 'first test value',
+        //     },
+        //     {
+        //       name: 'second',
+        //     },
+        //     {
+        //       name: 'third',
+        //       default: 'third-default',
+        //     },
+        //     {
+        //       name: 'fourth',
+        //       default: 'fourth-default',
+        //     },
+        //   ],
+        // });
+        jest.mocked(fetchBind).mockResolvedValue({
           provider: mockProviderJsonWithParameters,
           mapAst: mockMapDocument,
         });
@@ -208,7 +252,7 @@ describe('profile provider', () => {
         const mockProviderJsonWithParameters = mockProviderJson({
           name: 'test',
         });
-        mocked(fetchBind).mockResolvedValue({
+        jest.mocked(fetchBind).mockResolvedValue({
           provider: mockProviderJsonWithParameters,
           mapAst: mockMapDocument,
         });
@@ -260,7 +304,7 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -304,7 +348,7 @@ describe('profile provider', () => {
           cache: false,
         });
 
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -344,7 +388,7 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider use profile id', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -386,7 +430,7 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider use profile configuration', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -428,7 +472,7 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider load locally', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -476,7 +520,7 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider load localy and use map variant', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         fileSystem.readFile = jest
           .fn()
           .mockResolvedValueOnce(ok(JSON.stringify(providerJson)))
@@ -522,7 +566,7 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider with passed map variant and revision', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -563,7 +607,7 @@ describe('profile provider', () => {
         const result = await mockProfileProvider.bind();
 
         expect(result).toMatchObject(expectedBoundProfileProvider);
-        expect(fetchBind).toBeCalledWith(
+        expect(fetchBind).toHaveBeenCalledWith(
           {
             profileId: 'test-profile@1.0.0',
             provider: 'test',
@@ -578,7 +622,7 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider when map is provided locally but provider is not', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -605,7 +649,7 @@ describe('profile provider', () => {
           .fn()
           .mockResolvedValueOnce(ok(JSON.stringify(mockMapDocument)));
 
-        mocked(fetchProviderInfo).mockResolvedValue(providerJson);
+        jest.mocked(fetchProviderInfo).mockResolvedValue(providerJson);
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -626,7 +670,7 @@ describe('profile provider', () => {
       });
 
       it('loads provider from cache when fetch fails', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -658,7 +702,7 @@ describe('profile provider', () => {
           .mockResolvedValueOnce(ok(JSON.stringify(mockMapDocument)))
           .mockResolvedValueOnce(ok(JSON.stringify(providerJson)));
 
-        mocked(fetchProviderInfo).mockRejectedValue('denied!');
+        jest.mocked(fetchProviderInfo).mockRejectedValue('denied!');
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -678,7 +722,7 @@ describe('profile provider', () => {
       });
 
       it('throws when both fetch and cache read fails', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -708,7 +752,7 @@ describe('profile provider', () => {
           .mockResolvedValueOnce(ok(JSON.stringify(mockMapDocument)))
           .mockRejectedValueOnce(err('denied!'));
 
-        mocked(fetchProviderInfo).mockRejectedValue('denied!');
+        jest.mocked(fetchProviderInfo).mockRejectedValue('denied!');
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -776,7 +820,7 @@ describe('profile provider', () => {
       });
 
       it('throws error without profile provider settings', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -796,7 +840,8 @@ describe('profile provider', () => {
           new MockEnvironment()
         );
 
-        mocked(fileSystem.readFile)
+        jest
+          .mocked(fileSystem.readFile)
           .mockResolvedValueOnce(ok(JSON.stringify(mockProfileDocument)))
           .mockResolvedValueOnce(ok(JSON.stringify(mockProviderJson)))
           .mockResolvedValueOnce(ok(JSON.stringify(mockMapDocument)));
@@ -818,7 +863,7 @@ describe('profile provider', () => {
       });
 
       it('throws error when bind response contains invalid map AST', async () => {
-        mocked(fetchBind).mockResolvedValue({
+        jest.mocked(fetchBind).mockResolvedValue({
           ...mockFetchResponse,
           mapAst: undefined,
         });
@@ -843,7 +888,8 @@ describe('profile provider', () => {
           new MockEnvironment()
         );
 
-        mocked(fileSystem.readFile)
+        jest
+          .mocked(fileSystem.readFile)
           .mockResolvedValueOnce(ok(JSON.stringify(mockProfileDocument)))
           .mockResolvedValueOnce(ok(JSON.stringify(mockProviderJson)));
 
@@ -885,7 +931,7 @@ describe('profile provider', () => {
             password: 'test-digest-password',
           },
         ]);
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         const superJson = normalizeSuperJsonDocument(
           {
@@ -950,7 +996,7 @@ describe('profile provider', () => {
           } as any,
           new MockEnvironment()
         );
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -996,7 +1042,7 @@ but a secret value was provided for security scheme: made-up-id`
           new MockEnvironment()
         );
 
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -1042,7 +1088,7 @@ but apiKey scheme requires: apikey`
           new MockEnvironment()
         );
 
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -1088,7 +1134,7 @@ but http scheme requires: username, password`
           new MockEnvironment()
         );
 
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -1134,7 +1180,7 @@ but http scheme requires: token`
           new MockEnvironment()
         );
 
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         const mockProfileProvider = new ProfileProvider(
           superJson,
@@ -1155,7 +1201,7 @@ but http scheme requires: digest`
       });
 
       it('throws when super.json and provider.json provider names do not match', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
         const superJson = normalizeSuperJsonDocument(
           {
             profiles: {
@@ -1180,7 +1226,7 @@ but http scheme requires: digest`
           new MockEnvironment()
         );
 
-        mocked(fileSystem.readFile)
+        jest.mocked(fileSystem.readFile)
           .mockResolvedValueOnce(ok(JSON.stringify(providerJson)))
           .mockResolvedValueOnce(ok(JSON.stringify(mockMapDocument)));
 
@@ -1207,7 +1253,7 @@ but http scheme requires: digest`
       });
 
       it('throws when super.json and map provider names do not match', async () => {
-        mocked(fetchBind).mockResolvedValue(mockFetchResponse);
+        jest.mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         const superJson = normalizeSuperJsonDocument(
           {