+++ src/node/filesystem/index.ts
@@ -1,1 +1,3 @@
 export * from './filesystem.node';
+export { BinaryData } from './binary.node';
+export type { BinaryDataOptions } from './binary.node';