+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `BinaryData` type, not surfaced in Comlink Profile - [#296](https://github.com/superfaceai/one-sdk-js/pull/296), [#304](https://github.com/superfaceai/one-sdk-js/pull/304)
+
+## [2.1.0] - 2022-11-16
+### Added
+- `multipart/form-data` supports array values to define duplicate fields
+
+### Fixed
+- Do not set `Accept` and `Content-Type` header when defined in [HTTP request headers](https://spec.superface.dev/latest/map-spec.html#HTTPHeaders) - [#264](https://github.com/superfaceai/one-sdk-js/issues/264)
+- Replaced `isomorphic-form-data` with `form-data` package to fix `FormData` serialization - [#291](https://github.com/superfaceai/one-sdk-js/issues/291)
+- Create valid `headersInit` shape in `NodeFetch.fetch`
+- An http call to `/` url returning an error about it being an absolute url
 
 ## [2.0.0] - 2022-08-15
 ### Added
@@ -341,7 +353,8 @@ This is a recovery release after introduction of compatibility-breaking changes 
 - Profile parameter validator
 - CI/CD workflows
 
-[Unreleased]: https://github.com/superfaceai/one-sdk-js/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/superfaceai/one-sdk-js/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/superfaceai/one-sdk-js/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/superfaceai/one-sdk-js/compare/v1.5.2...v2.0.0
 [1.5.2]: https://github.com/superfaceai/one-sdk-js/compare/v1.5.1...v1.5.2
 [1.5.1]: https://github.com/superfaceai/one-sdk-js/compare/v1.5.0...v1.5.1